requirements.txt
requests==2.32.*
apscheduler==3.10.*
weasyprint==62.*
jinja2==3.1.*
python-dotenv==1.0.*
